requirements.txt
# Tkinter 为 Python 标准库，一般无需额外依赖
python-docx
streamlit
